Normalize floating-point literals to match Gappa's parser (fix issue #1).
This commit ensures that literals are normalized, so that Gappa cannot
modify them in any of the following ways.

Gappa normalizes floating-point literals on input, so that the mantissa is
not divisible by the radix. Moreover, if the exponent is zero after
normalization, the literal is turned into an integer. Also, zero is always
an integer. Finally, there are no negative floating-point literals, only
opposites of positive ones.