Fall back to plain ocamlfind as a last resort.
If Coq was compiled on MinGW but the "which" command is provided by
Cygwin, the latter will not make sense of the configuration of the former.
So, rather than continuing with an empty variable (which will later cause
strange error messages about option "-package" not being recognized), we
set it to "ocamlfind".